pyproject.toml
[project]
name = "src"
version = "2024.12.15"
authors = [
  {name="darrida", email="darrida.py@gmail.com"}
]
description = "CLI utility to help deploy py-shinylive apps to hosted webservers via SSH."
readme = "README.md"
requires-python = ">= 3.12"

dependencies = [
    "loguru",
    'logfire[django,asgi] > 2.0',
    # 'pydantic > 2.10',
    "django>=5.2.1",
    "social-auth-app-django>=5.4.3",
    "django-htmx>=1.21.0",
    "diskcache>=5.6.3",
    "onepassword-sdk>=0.1.5",
    "httpx>=0.28.1",
    "django-waffle>=4.2.0",
    "uvicorn>=0.34.0",
    "django-ninja>=1.4.1",
    "starlette>=0.46.2",
    "django-tasks>=0.7.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.25.0",
    "pytest-django>=4.9.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.logfire]
pydantic_plugin_record = "all"

[tool.ruff]
line-length = 120
extend-exclude = [".venv", "*.robot"]

[tool.ruff.lint]
extend-select = ["S", "I", "C4"]  # bandit (S), isort (I)
fixable = ["F401", "I001"]

[tool.ruff.lint.isort]
extra-standard-library = []
known-first-party = []
known-third-party = []

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "_core.settings"
# -- recommended but optional:
python_files = ["test_*.py", "_core/tests.py"]
